requirements.txt
# Core
pydantic>=2.10
pandas
pyarrow
httpx

# LLM
openai>=1.0.0
ollama

# Memory
chromadb>=0.5
sentence-transformers>=3.3
google-genai

# Infrastructure
apscheduler
python-dotenv

# Visualisation
plotly
